Integrate vec_iterator example into guide example
